Overall framework update
First of all, the Library folder contains several external libraries, such as imgui, MinHook, JSON, and font (used to draw icon text), the future may be added to the external libraries; Hook folder under the two files, used to Hook the game; Control folder under the file, used to write the code of the self-painted controls, I think that the menu, if you want to I think if the menu is to be beautiful, just relying on imgui's native controls is not enough, so it needs to be self-painted; Menu folder, is injected into the game menu GUI, which will be transported to the renderer for rendering, MainMenu file is not a sub-file of the Menu, but if I want to add a new sub-menu, I need to create another folder under the Menu folder. folder, e.g.. /Menu/Aimbot/Aimbot.cpp or .h; the Renderer folder is responsible for rendering, I want it to be a cool GUI, not a clichéd imgui GUI; Common.h and DllMain.cpp don't belong to any file, they belong to the NiceBase project. Common file contains almost all the header files, so in writing a new file, only need to include this file in the header file on the line, it is more convenient, at the same time, there are several define, not as good as NAME is “NiceBase”, PATH is the path under appdata, DllMain is the entrance to the DLL!